+++ lib/src/generators/common/resolve_asset_uri.dart
@@ -71,7 +71,8 @@ Future<Map<InterfaceType, String>> resolveAssetUris(
   for (final type in dartTypes) {
     final element = type.element;
     final elementLibrary = element.library;
-    if (elementLibrary.isInSdk && !elementLibrary.name.startsWith('dart._')) {
+    final elementLibraryName = elementLibrary.name ?? '';
+    if (elementLibrary.isInSdk && !elementLibraryName.startsWith('dart._')) {
       // For public SDK libraries, just use the source URI.
       typeUris[type] = elementLibrary.source.uri.toString();
       continue;