+++ lib/src/generators/mocker_generator.dart
@@ -1,6 +1,4 @@
-import 'package:analyzer/dart/constant/value.dart';
 import 'package:analyzer/dart/element/element.dart';
-import 'package:analyzer/dart/element/type.dart';
 import 'package:build/build.dart';
 import 'package:mockor/annotations.dart';
 import 'package:mockor/src/dartbuilders/mockor_dart_builder.dart';