+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Improve package description
 - Move builder.dart to src folder.
+- Fix compatibility with analyzer 1.7.0
 
 ## 1.0.0